+++ kit/src/Runtime/Math.c
@@ -20,26 +20,26 @@ max(unsigned int a, unsigned int b)
  *                         ML Integer Functions                           *
  *------------------------------------------------------------------------*/
 
-int 
-max_fixed_int(int dummy)                             /* ML */
+ssize_t 
+max_fixed_int(ssize_t dummy)                             /* ML */
 {
   return convertIntToML(Max_Int);
 }
 
-int 
-min_fixed_int(int dummy)                             /* ML */
+ssize_t 
+min_fixed_int(ssize_t dummy)                             /* ML */
 {
   return convertIntToML(Min_Int);
 }
 
-int 
-precision(int dummy)                                 /* ML */
+ssize_t 
+precision(ssize_t dummy)                                 /* ML */
 {
   return convertIntToML(val_precision);
 }
 
-int 
-__div_int31(int x, int y, int exn)                   /* ML */
+ssize_t 
+__div_int31(ssize_t x, ssize_t y, uintptr_t exn)                   /* ML */
 {
   if (y == 1) 
     { 
@@ -48,7 +48,7 @@ __div_int31(int x, int y, int exn)                   /* ML */
     }
   if ( y == -1 && x == -2147483647 )    // -2147483647 = 2 * Int31.minInt + 1
     {   
-      raise_exn((int)&exn_OVERFLOW);
+      raise_exn((uintptr_t)&exn_OVERFLOW);
       return 0;                         // never reached
     }
   if (x == 1) return 1;
@@ -59,8 +59,8 @@ __div_int31(int x, int y, int exn)                   /* ML */
   else return (2*((x-1)/(y-1))+1);
 }
 
-int 
-__div_int32ub(int x, int y, int exn)                 /* ML */
+ssize_t 
+__div_int32ub(ssize_t x, ssize_t y, uintptr_t exn)                 /* ML */
 {
   if (y == 0) 
     {
@@ -69,7 +69,7 @@ __div_int32ub(int x, int y, int exn)                 /* ML */
     }
   if ( y == -1 && x == (-2147483647 - 1) ) 
     {
-      raise_exn((int)&exn_OVERFLOW);
+      raise_exn((uintptr_t)&exn_OVERFLOW);
       return 0;                                // never reached
     }
   if (x < 0 && y > 0)
@@ -79,8 +79,8 @@ __div_int32ub(int x, int y, int exn)                 /* ML */
   else return x / y;
 }
 
-unsigned int 
-__div_word32ub(unsigned int x, unsigned int y, int exn)          /* ML */
+size_t 
+__div_word32ub(size_t x, size_t y, uintptr_t exn)          /* ML */
 {
   if ( y == 0 ) 
     {
@@ -90,11 +90,11 @@ __div_word32ub(unsigned int x, unsigned int y, int exn)          /* ML */
   return (x / y);
 }
 
-unsigned int 
-__div_word31(unsigned int x, unsigned int y, int exn)            /* ML */
+size_t 
+__div_word31(size_t x, size_t y, uintptr_t exn)            /* ML */
 {
-  unsigned int xC = i31_to_i32ub(x);
-  unsigned int yC = i31_to_i32ub(y);
+  size_t xC = i31_to_i32ub(x);
+  size_t yC = i31_to_i32ub(y);
 
   if ( yC == 0 ) 
     {
@@ -104,8 +104,8 @@ __div_word31(unsigned int x, unsigned int y, int exn)            /* ML */
   return i32ub_to_i31(xC / yC);
 }
 
-int 
-__mod_int31(int xML, int yML, int exn) 
+ssize_t 
+__mod_int31(ssize_t xML, ssize_t yML, uintptr_t exn) 
 {
   if ( yML == 1 ) 
     {
@@ -118,8 +118,8 @@ __mod_int31(int xML, int yML, int exn)
     return ((xML-1)%(yML-1))+yML;
 }
 
-int 
-__mod_int32ub(int x, int y, int exn) 
+ssize_t 
+__mod_int32ub(ssize_t x, ssize_t y, uintptr_t exn) 
 {
   if ( y == 0 ) 
     {
@@ -133,8 +133,8 @@ __mod_int32ub(int x, int y, int exn)
   return (x % y) + y;
 }
 
-unsigned int 
-__mod_word32ub(unsigned int x, unsigned int y, int exn) 
+size_t 
+__mod_word32ub(size_t x, size_t y, uintptr_t exn) 
 {
   if ( y == 0 ) 
     {
@@ -144,11 +144,11 @@ __mod_word32ub(unsigned int x, unsigned int y, int exn)
   return (x % y);
 }
 
-unsigned int 
-__mod_word31(unsigned int x, unsigned int y, int exn) 
+size_t 
+__mod_word31(size_t x, size_t y, uintptr_t exn) 
 {
-  unsigned int xC = i31_to_i32ub(x);
-  unsigned int yC = i31_to_i32ub(y);
+  size_t xC = i31_to_i32ub(x);
+  size_t yC = i31_to_i32ub(y);
 
   if ( yC == 0 ) 
     {
@@ -158,32 +158,32 @@ __mod_word31(unsigned int x, unsigned int y, int exn)
   return i32ub_to_i31(xC % yC);
 }
 
-int 
-__quot_int32ub(int xML, int yML) 
+ssize_t 
+__quot_int32ub(ssize_t xML, ssize_t yML) 
 {
   return xML/yML;
 }
 
-int 
-__quot_int31(int xML, int yML) 
+ssize_t 
+__quot_int31(ssize_t xML, ssize_t yML) 
 {
-  int xC,yC;
+  ssize_t xC,yC;
 
   xC = i31_to_i32ub(xML);
   yC = i31_to_i32ub(yML);
   return i32ub_to_i31(xC / yC);
 }
 
-int 
-__rem_int32ub(int xML, int yML) 
+ssize_t 
+__rem_int32ub(ssize_t xML, ssize_t yML) 
 {
   return xML % yML;
 }
 
-int 
-__rem_int31(int xML, int yML) 
+ssize_t 
+__rem_int31(ssize_t xML, ssize_t yML) 
 {
-  int xC,yC;
+  ssize_t xC,yC;
 
   xC = i31_to_i32ub(xML);
   yC = i31_to_i32ub(yML);
@@ -193,32 +193,32 @@ __rem_int31(int xML, int yML)
 
 #ifdef TAG_VALUES
 
-unsigned int* 
-__div_int32b(unsigned int* b, unsigned int* x, unsigned int* y, int exn) 
+size_t* 
+__div_int32b(size_t* b, size_t* x, size_t* y, uintptr_t exn) 
 {
   get_i32b(b) = __div_int32ub(get_i32b(x), get_i32b(y), exn);
   set_i32b_tag(b);
   return b;
 }
 
-unsigned int* 
-__div_word32b(unsigned int* b, unsigned int* x, unsigned int* y, int exn) 
+size_t* 
+__div_word32b(size_t* b, size_t* x, size_t* y, uintptr_t exn) 
 {
   get_i32b(b) = __div_word32ub(get_i32b(x), get_i32b(y), exn);
   set_i32b_tag(b);
   return b;
 }
 
-unsigned int* 
-__mod_int32b(unsigned int* b, unsigned int* x, unsigned int* y, int exn) 
+size_t* 
+__mod_int32b(size_t* b, size_t* x, size_t* y, uintptr_t exn) 
 {
   get_i32b(b) = __mod_int32ub(get_i32b(x), get_i32b(y), exn);
   set_i32b_tag(b);
   return b;
 }
 
-unsigned int* 
-__mod_word32b(unsigned int* b, unsigned int* x, unsigned int* y, int exn) 
+size_t* 
+__mod_word32b(size_t* b, size_t* x, size_t* y, uintptr_t exn) 
 {
   get_i32b(b) = __mod_word32ub(get_i32b(x), get_i32b(y), exn);
   set_i32b_tag(b);
@@ -226,8 +226,8 @@ __mod_word32b(unsigned int* b, unsigned int* x, unsigned int* y, int exn)
 }
 
 // quot need not check for y being 0; this is checked for in Int32
-unsigned int* 
-__quot_int32b(unsigned int* b, unsigned int* x, unsigned int* y) 
+size_t* 
+__quot_int32b(size_t* b, size_t* x, size_t* y) 
 {
   get_i32b(b) = __quot_int32ub(get_i32b(x), get_i32b(y));
   set_i32b_tag(b);
@@ -235,8 +235,8 @@ __quot_int32b(unsigned int* b, unsigned int* x, unsigned int* y)
 }
 
 // rem need not check for y being 0; this is checked for in Int32
-unsigned int* 
-__rem_int32b(unsigned int* b, unsigned int* x, unsigned int* y) 
+size_t* 
+__rem_int32b(size_t* b, size_t* x, size_t* y) 
 {
   get_i32b(b) = __rem_int32ub(get_i32b(x), get_i32b(y));
   set_i32b_tag(b);
@@ -245,8 +245,8 @@ __rem_int32b(unsigned int* b, unsigned int* x, unsigned int* y)
 
 #endif /*TAG_VALUES*/
 
-int 
-realInt(int d, int x) 
+ssize_t 
+realInt(ssize_t d, ssize_t x) 
 {
   debug(printf("[realInt: d = %x, x = %d\n", d, x));
   get_d(d) = (double) (convertIntToC(x));
@@ -259,34 +259,34 @@ realInt(int d, int x)
  *                      Floating point operations                       *
  *----------------------------------------------------------------------*/
 
-int 
-divFloat(int d, int x, int y) 
+ssize_t 
+divFloat(ssize_t d, ssize_t x, ssize_t y) 
 {
   get_d(d) = get_d(x) / get_d(y);
   set_dtag(d);
   return d;
 }
 
-int 
-floorFloat(int f) 
+ssize_t 
+floorFloat(ssize_t f) 
 { 
   double r;
-  int i;
+  ssize_t i;
 
   r = get_d(f);
   if( r >= 0.0 ) 
     { 
       if ( r >= (Max_Int_d + 1.0) ) 
 	{
-	  raise_exn((int)&exn_OVERFLOW);
+	  raise_exn((uintptr_t)&exn_OVERFLOW);
 	}
-      return (convertIntToML((int) r));
+      return (convertIntToML((ssize_t) r));
     }
   if( r < Min_Int_d ) 
     {
-      raise_exn((int)&exn_OVERFLOW);
+      raise_exn((uintptr_t)&exn_OVERFLOW);
     }
-  i = (int) r;
+  i = (ssize_t) r;
   if( r < ((double) i) ) 
     {
       i -= 1;
@@ -294,150 +294,150 @@ floorFloat(int f)
   return convertIntToML(i);
 }
 
-int 
-truncFloat(int f) 
+ssize_t 
+truncFloat(ssize_t f) 
 { 
   double r;
 
   r = get_d(f);
   if ((r >= (Max_Int_d + 1.0)) || (r <= (Min_Int_d - 1.0))) 
     { 
-      raise_exn((int)&exn_OVERFLOW);
+      raise_exn((uintptr_t)&exn_OVERFLOW);
     }
-  return convertIntToML((int)r);
+  return convertIntToML((ssize_t)r);
 }
 
-int 
-ceilFloat(int f) 
+ssize_t 
+ceilFloat(ssize_t f) 
 { 
   double arg;
-  int i;
+  ssize_t i;
 
   arg = get_d(f);
 
   if( arg >= 0.0 ) 
     { 
       if( arg > Max_Int_d ) goto raise_ceil;
-      i = (int) arg;
+      i = (ssize_t) arg;
       if( arg > ((double) i) ) i += 1;
     }
   else 
     { 
       if( arg <= (Min_Int_d - 1.0) ) goto raise_ceil;
-      i = (int) arg;
+      i = (ssize_t) arg;
     }
   return convertIntToML(i);
 
  raise_ceil:
-  raise_exn((int)&exn_OVERFLOW);
+  raise_exn((uintptr_t)&exn_OVERFLOW);
   return 0;                          // never reached
 }
 
-int 
-sqrtFloat(int d, int s) 
+ssize_t 
+sqrtFloat(ssize_t d, ssize_t s) 
 { 
   get_d(d) = sqrt(get_d(s)); 
   set_dtag(d);
   return d;
 }
 
-int 
-sinFloat(int d, int s) 
+ssize_t 
+sinFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = sin(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-cosFloat(int d, int s) 
+ssize_t 
+cosFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = cos(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-atanFloat (int d, int s) 
+ssize_t 
+atanFloat (ssize_t d, ssize_t s) 
 {
   get_d (d) = atan (get_d (s));
   set_dtag(d);
   return d;
 }
 
-int 
-asinFloat (int d, int s) 
+ssize_t 
+asinFloat (ssize_t d, ssize_t s) 
 {
   get_d (d) = asin (get_d (s));
   set_dtag(d);
   return d;
 }
 
-int 
-acosFloat (int d, int s) 
+ssize_t 
+acosFloat (ssize_t d, ssize_t s) 
 {
   get_d (d) = acos (get_d (s));
   set_dtag(d);
   return d;
 }
 
-int 
-atan2Float (int d, int y, int x) 
+ssize_t 
+atan2Float (ssize_t d, ssize_t y, ssize_t x) 
 {
   get_d (d) = atan2 (get_d (y), get_d (x));
   set_dtag(d);
   return d;
 }
 
-int expFloat(int d, int s) 
+ssize_t expFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = exp(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-powFloat (int d, int x, int y) 
+ssize_t 
+powFloat (ssize_t d, ssize_t x, ssize_t y) 
 {
   get_d (d) = pow (get_d (x), get_d (y));
   set_dtag(d);
   return d;
 }
 
-int 
-lnFloat(int d, int s) 
+ssize_t 
+lnFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = log(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-sinhFloat(int d, int s) 
+ssize_t 
+sinhFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = sinh(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-coshFloat(int d, int s) 
+ssize_t 
+coshFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = cosh(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-tanhFloat(int d, int s) 
+ssize_t 
+tanhFloat(ssize_t d, ssize_t s) 
 {
   get_d(d) = tanh(get_d(s));
   set_dtag(d);
   return d;
 }
 
-int 
-isnanFloat(int s) 
+ssize_t 
+isnanFloat(ssize_t s) 
 {
   if (isnan(get_d(s))) 
     {
@@ -446,16 +446,16 @@ isnanFloat(int s)
   return mlFALSE;
 }
 
-int 
-posInfFloat(int d) 
+ssize_t 
+posInfFloat(ssize_t d) 
 {
   get_d(d) = HUGE_VAL;
   set_dtag(d);
   return d;
 }
 
-int 
-negInfFloat(int d) 
+ssize_t 
+negInfFloat(ssize_t d) 
 {
   get_d(d) = -HUGE_VAL;
   set_dtag(d);
@@ -464,9 +464,9 @@ negInfFloat(int d)
 
 // countChar: count the number of times the character `c' 
 //     occurs in the string `s'.
-static int countChar(int c, char * s) {
+static ssize_t countChar(ssize_t c, char * s) {
   char *p; 
-  int count;
+  ssize_t count;
 
   count = 0;
   for( p=s; *p != '\0'; p++ ) 
@@ -494,7 +494,7 @@ static void mkSMLMinus(char * s) {
 }
 
 String
-REG_POLY_FUN_HDR(stringOfFloat, Region rAddr, int arg) 
+REG_POLY_FUN_HDR(stringOfFloat, Region rAddr, size_t arg) 
 {
   char buf[64];
   sprintf(buf, "%.12g", get_d(arg));
@@ -507,7 +507,7 @@ REG_POLY_FUN_HDR(stringOfFloat, Region rAddr, int arg)
 }
 
 String
-REG_POLY_FUN_HDR(generalStringOfFloat, Region rAddr, String format, int f) 
+REG_POLY_FUN_HDR(generalStringOfFloat, Region rAddr, String format, size_t f) 
 {
   char result_buf[512];
 
@@ -522,16 +522,16 @@ REG_POLY_FUN_HDR(generalStringOfFloat, Region rAddr, String format, int f)
 
 /* DEBUG */
 void 
-printReal(int f) 
+printReal(size_t f) 
 {
   printf("Num: %5.2f\n",get_d(f));
   return;
 }
 
 String
-REG_POLY_FUN_HDR(sml_real_to_bytes, Region rAddr, int f)
+REG_POLY_FUN_HDR(sml_real_to_bytes, Region rAddr, size_t f)
 {
-  int i;
+  size_t i;
   double a[1];
   char v[sizeof(double) + 1];
   a[0] = get_d(f);
@@ -541,8 +541,8 @@ REG_POLY_FUN_HDR(sml_real_to_bytes, Region rAddr, int f)
   return REG_POLY_CALL(convertBinStringToML, rAddr, 8, v);
 }
 
-int 
-sml_bytes_to_real(int d, String s)
+size_t 
+sml_bytes_to_real(size_t d, String s)
 {
   double r;
   char* a = &(s->data);
